Added few Repositories , Services and Controllers Build a basic CLI UI for Payment Management System Todo - testing , code coverage , adding Diagrams , Readme

+++ src/main/java/com/payments/model/Role.java
@@ -1,23 +1,20 @@
 package com.payments.model;
 
+//new
+
 public class Role {
-    private Long id;
-    private String name; // ADMIN, FINANCE_MANAGER, VIEWER
+    private Integer id;
+    private String name;
 
     public Role() {}
-    public Role(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
+    public Role(Integer id, String name) { this.id = id; this.name = name; }
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public Integer getId() { return id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
 
     @Override
-    public String toString() {
-        return "Role{id=" + id + ", name='" + name + "'}";
-    }
+    public String toString() { return "Role{id=" + id + ", name='" + name + "'}"; }
 }
